+++ src/test/resources/cuke/choonz.feature
@@ -1,6 +1,6 @@
 Feature: Choonz website tests
 
-  Scenario: As a User I want to create an account on the website so that I can login
+  Scenario Outline: As a User I want to create an account on the website so that I can login
     Given that I can navigate to "http://localhost:8082/index.html"
     When I click the account button
     And I cancel the alert
@@ -12,3 +12,4 @@ Feature: Choonz website tests
 		Examples:
   		| username | password |
   		| test_user | test_password |
+
